fix(desktop): isolate notification event mapping for tests
